Added Landing and Room page
