+++ src/App/Index.tsx
@@ -11,6 +11,7 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 import Home from "../pages/Home";
 import Login from "../pages/Login";
 import Register from "../pages/Register";
+import Room from "../pages/Room";
 import Navbar from "../components/Navbar";
 import UserProvider from "../store/user";
 import theme from "../theme";
@@ -28,6 +29,7 @@ const App: React.FC = () => {
             <div className="App">
               <Switch>
                 <Route path="/" exact component={Home} />
+                <Route path="/room" component={Room} />
                 <Route path="/login" component={Login} />
                 <Route path="/register" component={Register} />
                 <Redirect to="/" />
