+++ src/pages/Home/Home.module.scss
@@ -1,15 +1,9 @@
-.ytPlayer {
-  width: 100%;
-  height: 60vh;
-}
-
 .HomeContainer {
   width: 100%;
   margin-top: 30px;
   height: 60vh;
-}
-
-.NextLabel {
-  display: flex;
-  margin: 20px;
+  display:flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
 }
